modules/accelerators/contextualization/cdf_three_dimenssion/functions/fn_context_3d_cad_asset_contextualization/requirements.txt
cognite-extractor-utils>=7
cognite-sdk>=7.26
pyyaml>=6
requests>=2.28
